+++ packages/create-skeleton-app/templates/skeleton-template-appshell/src/routes/+layout.svelte
@@ -13,7 +13,14 @@
 			</svelte:fragment>
 			<svelte:fragment slot="trail">
 				<a class="btn btn-sm variant-ghost-surface" href="https://discord.gg/EXqV7W8MtY" target="_blank" rel="noreferrer"> Discord </a>
-				<a class="btn btn-sm variant-ghost-surface" href="https://twitter.com/SkeletonUI" target="_blank" rel="noreferrer"> Twitter </a>
+				<a
+					class="btn btn-sm variant-ghost-surface"
+					href="https://bsky.app/profile/skeletonlabs.bsky.social"
+					target="_blank"
+					rel="noreferrer"
+				>
+					Bluesky
+				</a>
 				<a class="btn btn-sm variant-ghost-surface" href="https://github.com/skeletonlabs/skeleton" target="_blank" rel="noreferrer">
 					GitHub
 				</a>