+++ packages/necroparser/test/fixtures/extends-external.ts
@@ -1,0 +1,6 @@
+import type { Context } from '@zag-js/accordion';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+interface Interface extends Context {
+	property: string;
+}