+++ packages/skeleton-svelte/src/lib/components/Popover/types.ts
@@ -7,33 +7,51 @@ export interface PopoverProps extends Omit<popover.Context, 'id' | 'open'> {
 	/** Enable display of the popover arrow. */
 	arrow?: boolean;
 
+	// Base ---
+	/** Set base classes for the root element. */
+	base?: string;
+	/** Provide arbitrary classes for the root element. */
+	classes?: string;
+
 	// Trigger ---
-	/** Set base styles for the trigger. */
+	/** Set base classes for the trigger. */
 	triggerBase?: string;
-	/** Set background styles for the trigger. */
+	/** Set background classes for the trigger. */
 	triggerBackground?: string;
-	/** Provide arbitrary styles for the trigger. */
+	/** Provide arbitrary classes for the trigger. */
 	triggerClasses?: string;
 
+	// Positioner ---
+	/** Set base classes for the positioner. */
+	positionerBase?: string;
+	/** Set z-index classes for the positioner. */
+	positionerZIndex?: string;
+	/** Provide arbitrary classes for the positioner. */
+	positionerClasses?: string;
+
 	// Content ---
-	/** Set base styles for the content. */
+	/** Set base classes for the content. */
 	contentBase?: string;
-	/** Set background styles for the content. */
+	/** Set background classes for the content. */
 	contentBackground?: string;
-	/** Provide arbitrary styles for the content. */
+	/** Provide arbitrary classes for the content. */
 	contentClasses?: string;
 
 	// Arrow ---
-	/** Set base styles for the arrow. */
+	/** Set base classes for the arrow. */
 	arrowBase?: string;
-	/** Set background styles for the arrow. */
+	/** Set background classes for the arrow. */
 	arrowBackground?: string;
-	/** Provide arbitrary styles for the arrow. */
+	/** Provide arbitrary classes for the arrow. */
 	arrowClasses?: string;
 
 	// Snippets ---
-	/** Provide the the trigger element, such as a button. */
+	/** Provide the template contents inside the trigger button. */
 	trigger?: Snippet;
 	/** Provide the template contents of the popover itself. */
 	content?: Snippet;
+
+	// Events ---
+	/** Handle the popover button click event. */
+	onclick?: () => void;
 }